fixup(p1b-bug1): split full_delete_instance + name_residual_anywhere + tests to instance_lifecycle.rs (file_size_invariant fix)
## Why
CI failed on `tests/file_size_invariant.rs::mcp_handler_files_under_500_loc` — the +290 LOC of Bug 1 fix pushed `src/mcp/handlers/instance.rs` from 677 LOC (just under the 700 ceiling) to 947 LOC (247 over). Lead approved Option 3 (split, not invariant bypass) per Sprint 54 fix-root-cause philosophy.

## What changed
1. **New `src/mcp/handlers/instance_lifecycle.rs`** (~344 LOC): hosts `full_delete_instance` + `name_residual_anywhere` + the 10 unit tests. Module doc explains the rationale (file_size_invariant ceiling) + public surface.
2. **`src/mcp/handlers/instance.rs`** trimmed from 947 → 631 LOC (well under 700 ceiling): removed the moved functions + test mod; `handle_delete_instance` now dispatches via `super::instance_lifecycle::full_delete_instance(...)`. Unused imports (`cleanup_working_dir`, `crate::channel::telegram`) cleaned up.
3. **`src/mcp/handlers/mod.rs`**: `pub(crate) mod instance_lifecycle;` declaration added next to existing `pub(crate) mod instance;`.
4. **`src/app/overlay.rs`** (TUI close paths): `crate::mcp::handlers::instance::full_delete_instance` → `crate::mcp::handlers::instance_lifecycle::full_delete_instance` (2 sites).

## Production behaviour unchanged
- Same `full_delete_instance` signature: `(home, name) -> Result<(), String>`
- Same `name_residual_anywhere` signature + 6-store enumeration
- `handle_delete_instance` MCP path unchanged: same JSON shape on Ok/Err
- TUI close paths unchanged: same `tracing::warn!` on Err
- 10 unit tests all pass at the new location (1641 lib total)

## Verification
- `cargo build` + `cargo fmt --check` + `cargo clippy --all-targets -- -D warnings`: clean
- `cargo test --lib mcp::handlers::instance`: 10 passed at new module path (`mcp::handlers::instance_lifecycle::tests::*`)
- `cargo test --test file_size_invariant`: 1 passed (was failing on b06d7ac)
- Full lib test: 1641 passed / 0 failed

## §3.5.13 push form scope-claim (post-r0 fixup addendum)
post-r0 fixup commit: split `full_delete_instance` + `name_residual_anywhere` + 10 tests to new sibling `src/mcp/handlers/instance_lifecycle.rs` (instance.rs 947→631 within 700 file_size_invariant ceiling, instance_lifecycle.rs ~344 LOC); legitimate module split per file_size_invariant test design intent (not invariant bypass); production behaviour unchanged; no caller chain impact (`handle_delete_instance` still in instance.rs as 1-line dispatch wrapper; `app/overlay.rs` 2 callers updated to new module path); future-proofs the file_size_invariant test for further `instance_lifecycle` work (e.g. P1-B Bug 2 wrapper UX fix when embargo lifts).

Decision/task: `t-20260507232635952928-12`. Reviewer m-20260507233728707015-240 had flagged Check 3 as PARTIAL on registry/externals post-audit nuance; the new module's `name_residual_anywhere` doc now explicitly addresses this (audit covers disk-backed stores where `auto_start_fleet` revival risk lives; in-process registry/externals are cleared by the DELETE step itself).
